# zstd: Minor decoder improvements